feat: js -> ts and split to modules
