+++ package.json
@@ -2,22 +2,26 @@
     "name": "ws_task",
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
-    "main": "index.js",
-    "type": "module",
+    "main": "src/index.ts",
     "scripts": {
-        "start": "node index.js",
-        "start:dev": "nodemon index.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
+        "start": "tsc && node dist/index.js",
+        "start:dev": "nodemon --exec npx ts-node-esm src/index.ts",
+        "kill": "kill -9 $(lsof -ti:3000)"
     },
     "author": "Andrei Auchynnikau",
     "keywords": [],
     "license": "ISC",
     "devDependencies": {
-        "nodemon": "^2.0.16"
+        "@types/jimp": "^0.2.28",
+        "@types/node": "^18.0.0",
+        "@types/ws": "^8.5.3",
+        "nodemon": "^2.0.16",
+        "ts-node": "^10.8.1"
     },
     "dependencies": {
         "jimp": "^0.16.1",
         "robotjs": "^0.6.0",
+        "typescript": "^4.7.4",
         "ws": "^8.8.0"
     }
 }
