+++ src/constants.ts
@@ -1,0 +1,13 @@
+export const directions = {
+    up: 'mouse_up',
+    down: 'mouse_down',
+    right: 'mouse_right',
+    left: 'mouse_left'
+}
+
+
+export const figures = {
+    circle: 'draw_circle',
+    rectangle: 'draw_rectangle',
+    square: 'draw_square'
+}
